requirements_inference.txt
# XGait Inference Pipeline Requirements
# Core dependencies for person identification using XGait model

# Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Object Detection
ultralytics>=8.0.0

# Numerical Computing
numpy>=1.21.0

# Utility Libraries
tqdm>=4.60.0

# Optional: For better performance on specific hardware
# torch-audio>=0.12.0  # For MPS support on Apple Silicon
# cupy>=10.0.0         # For CUDA acceleration (NVIDIA GPUs)

# Development/Testing (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0
